feat(assistant): finance tool via shared computeAnalytics; parity-pinned to /analytics
Extract the GET /api/analytics handler body into analyticsService.computeAnalytics so
both the route and the new financial_summary assistant tool call identical code — their
numbers are structurally guaranteed to match. Parity enforced by an integration test.
